+++ src/transport/subscription-worker/subscription-worker-middleware.ts
@@ -12,13 +12,13 @@ import { TransportResponse } from '../../core/types/transport-response';
 import * as PubNubSubscriptionWorker from './subscription-worker-types';
 import { LoggerManager } from '../../core/components/logger-manager';
 import { LogLevel, LogMessage } from '../../core/interfaces/logger';
+import { Status, StatusEvent, Payload } from '../../core/types/api';
 import { TokenManager } from '../../core/components/token_manager';
 import { RequestSendingError } from './subscription-worker-types';
 import { PubNubAPIError } from '../../errors/pubnub-api-error';
 import StatusCategory from '../../core/constants/categories';
 import { Transport } from '../../core/interfaces/transport';
 import * as PAM from '../../core/types/api/access-manager';
-import { Status, StatusEvent } from '../../core/types/api';
 import PNOperations from '../../core/constants/operations';
 
 // --------------------------------------------------------
@@ -181,6 +181,22 @@ export class SubscriptionWorkerMiddleware implements Transport {
       subscriptionKey: this.configuration.subscriptionKey,
       userId: this.configuration.userId,
       workerLogLevel: this.configuration.workerLogLevel,
+    });
+  }
+
+  /**
+   * Update presence state associated with `userId`.
+   *
+   * @param state - Key-value pair of payloads (states) that should be associated with channels / groups specified as
+   * keys.
+   */
+  onPresenceStateChange(state: Record<string, Payload>) {
+    this.scheduleEventPost({
+      type: 'client-presence-state-update',
+      clientIdentifier: this.configuration.clientIdentifier,
+      subscriptionKey: this.configuration.subscriptionKey,
+      workerLogLevel: this.configuration.workerLogLevel,
+      state,
     });
   }
 
@@ -510,7 +526,7 @@ export class SubscriptionWorkerMiddleware implements Transport {
       if (!token || !stringifiedToken) return undefined;
 
       return (this.accessTokensMap = {
-        [accessToken]: { token: stringifiedToken, expiration: token.timestamp * token.ttl * 60 },
+        [accessToken]: { token: stringifiedToken, expiration: token.timestamp + token.ttl * 60 },
       })[accessToken];
     });
   }